+++ e2e/__tests__/jest.config.ts.test.ts
@@ -176,6 +176,30 @@ onNodeVersions('^23.6', () => {
     ).toMatchSnapshot();
     expect(exitCode).toBe(1);
   });
+
+  test('load typed jest.config.ts with TS loader specified in docblock pragma', () => {
+    writeFiles(DIR, {
+      '__tests__/a-giraffe.js': "test('giraffe', () => expect(1).toBe(1));",
+      'foo.ts': 'export const a = () => {};',
+      'jest.config.ts':
+      `
+        /** @jest-config-loader ts-node */
+        import { a } from './foo'
+        a();
+        import type {Config} from 'jest';
+        const config: Config = { testTimeout: 10000 };
+        export default config;
+      `,
+      'package.json': '{}',
+    }); 
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=false'], {
+      nodeOptions: '--no-warnings',
+    });
+    const {rest, summary} = extractSummary(stderr);
+    expect(exitCode).toBe(0);
+    expect(rest).toMatchSnapshot();
+    expect(summary).toMatchSnapshot();
+  })
 });
 
 onNodeVersions('>=24', () => {